# sqle/docs/swagger.yaml
basePath: /
definitions:
  controller.BaseRes:
    properties:
      code:
        example: 0
        type: integer
      message:
        example: ok
        type: string
    type: object
...
  v1.CreateUserGroupReqV1:
    properties:
      is_disabled:
        type: boolean
      user_group_desc:
        example: this is a group
        type: string
      user_group_name:
        example: test
        type: string
      user_name_list:
        items:
          type: string
        type: array
    type: object
  v1.CreateUserReqV1:
    properties:
      email:
        example: test@email.com
        type: string
      role_name_list:
        items:
          type: string
        type: array
      user_name:
        example: test
        type: string
      user_password:
        example: "123456"
        type: string
    type: object
...
  v1.GetUserDetailResV1:
    properties:
      code:
        example: 0
        type: integer
      data:
        $ref: '#/definitions/v1.UserDetailResV1'
        type: object
      message:
        example: ok
        type: string
    type: object
  v1.GetUserGroupsResV1:
    properties:
      code:
        example: 0
        type: integer
      data:
        items:
          $ref: '#/definitions/v1.UserGroupListItemResV1'
        type: array
      message:
        example: ok
        type: string
      total_nums:
        type: integer
    type: object
...
  v1.GetUsersResV1:
    properties:
      code:
        example: 0
        type: integer
      data:
        items:
          $ref: '#/definitions/v1.UserResV1'
        type: array
      message:
        example: ok
        type: string
      total_nums:
        type: integer
    type: object
...
  v1.UpdateCurrentUserReqV1:
    properties:
      email:
        type: string
    type: object
...
  v1.UpdateUserReqV1:
    properties:
      email:
        type: string
      role_name_list:
        items:
          type: string
        type: array
    type: object
...
  v1.UserDetailResV1:
    properties:
      email:
        type: string
      is_admin:
        type: boolean
      login_type:
        type: string
      role_name_list:
        items:
          type: string
        type: array
      user_name:
        type: string
    type: object
  v1.UserGroupListItemResV1:
    properties:
      group_desc:
        type: string
      group_name:
        type: string
      id:
        type: integer
      is_disabled:
        type: boolean
    type: object
...
  v1.UserResV1:
    properties:
      email:
        type: string
      login_type:
        type: string
      role_name_list:
        items:
          type: string
        type: array
      user_name:
        type: string
    type: object
...
info:
  contact: {}
  description: This is a sample server for dev.
  license: {}
  title: Sqle API Docs
  version: "1.0"
paths:
...
  /v1/user:
    get:
      description: get current user info
      operationId: getCurrentUserV1
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.GetUserDetailResV1'
      security:
      - ApiKeyAuth: []
      summary: 获取当前用户信息
      tags:
      - user
    patch:
      consumes:
      - application/json
      description: update current user
      operationId: updateCurrentUserV1
      parameters:
      - description: update user
        in: body
        name: instance
        required: true
        schema:
          $ref: '#/definitions/v1.UpdateCurrentUserReqV1'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.BaseRes'
      security:
      - ApiKeyAuth: []
      summary: 更新个人信息
      tags:
      - user
...
  /v1/user_groups:
    get:
      description: get user group info list
      operationId: getUserGroupListV1
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.GetUserGroupsResV1'
      security:
      - ApiKeyAuth: []
      summary: 获取用户组列表
      tags:
      - user_group
    post:
      consumes:
      - application/json
      description: create user group
      operationId: CreateUserGroupV1
      parameters:
      - description: create user group
        in: body
        name: instance
        required: true
        schema:
          $ref: '#/definitions/v1.CreateUserGroupReqV1'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.BaseRes'
      security:
      - ApiKeyAuth: []
      summary: 创建用户组
      tags:
      - user_group
  /v1/user_groups/{user_group_name}/:
    delete:
      description: delete user group
      operationId: deleteUserV1
      parameters:
      - description: user_group_name
        in: path
        name: user_group_name
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.BaseRes'
      security:
      - ApiKeyAuth: []
      summary: 删除用户组
      tags:
      - user_group
...
  /v1/users:
    get:
      description: get user info list
      operationId: getUserListV1
      parameters:
      - description: filter user name
        in: query
        name: filter_user_name
        type: string
      - description: filter role name
        in: query
        name: filter_role_name
        type: string
      - description: page index
        in: query
        name: page_index
        type: integer
      - description: size of per page
        in: query
        name: page_size
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.GetUsersResV1'
      security:
      - ApiKeyAuth: []
      summary: 获取用户信息列表
      tags:
      - user
    post:
      consumes:
      - application/json
      description: create user
      operationId: createUserV1
      parameters:
      - description: create user
        in: body
        name: instance
        required: true
        schema:
          $ref: '#/definitions/v1.CreateUserReqV1'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.BaseRes'
      security:
      - ApiKeyAuth: []
      summary: 创建用户
      tags:
      - user
  /v1/users/{user_name}/:
    delete:
      description: delete user
      operationId: deleteUserV1
      parameters:
      - description: user name
        in: path
        name: user_name
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.BaseRes'
      security:
      - ApiKeyAuth: []
      summary: 删除用户
      tags:
      - user
    get:
      description: get user info
      operationId: getUserV1
      parameters:
      - description: user name
        in: path
        name: user_name
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.GetUserDetailResV1'
      security:
      - ApiKeyAuth: []
      summary: 获取用户信息
      tags:
      - user
    patch:
      consumes:
      - application/json
      description: update user
      operationId: updateUserV1
      parameters:
      - description: user name
        in: path
        name: user_name
        required: true
        type: string
      - description: update user
        in: body
        name: instance
        required: true
        schema:
          $ref: '#/definitions/v1.UpdateUserReqV1'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.BaseRes'
      security:
      - ApiKeyAuth: []
      summary: 更新用户信息
      tags:
      - user
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"